프로젝트, 회사 crud

+++ src/main/java/com/lunchapp/service/project/ProjectService.java
@@ -1,18 +1,24 @@
 package com.lunchapp.service.project;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.lunchapp.constants.Constants;
+import com.lunchapp.constants.DBNAME;
 import com.lunchapp.exception.DuplicateException;
 import com.lunchapp.exception.NoSearchObjectException;
+import com.lunchapp.model.dto.BaseDto;
 import com.lunchapp.model.dto.Result;
-import com.lunchapp.model.dto.todolist.ProjectDto;
+import com.lunchapp.model.dto.project.ProjectDto;
 import com.lunchapp.model.project.Company;
 import com.lunchapp.model.project.CompanyRepository;
+import com.lunchapp.model.project.PSTATUS;
 import com.lunchapp.model.project.Project;
 import com.lunchapp.model.project.ProjectRepository;
 
@@ -32,7 +38,7 @@ public class ProjectService {
 			Project project = new Project(dto);
 			if (dto.getCompanyId() != null) {
 				Company company = companyRepository.findById(Long.valueOf(dto.getCompanyId()))
-						.orElseThrow(() -> new NoSearchObjectException("회사"));
+						.orElseThrow(() -> new NoSearchObjectException(DBNAME.PROJECT.getKrname()));
 				project.setCompany(company);
 			}
 			projectRepository.save(project);
@@ -46,7 +52,10 @@ public class ProjectService {
 	public Result<List<ProjectDto>> deleteProject(ProjectDto dto) {
 		try {
 			Project project = projectRepository.findById(Long.valueOf(dto.getId()))
-					.orElseThrow(() -> new NoSearchObjectException("프로젝트"));
+					.orElseThrow(() -> new NoSearchObjectException(DBNAME.PROJECT.getKrname()));
+			if (project.getProjectStatus() == PSTATUS.PROCEEDING)
+				throw new Exception("진행중인 프로젝트는 삭제할 수 없습니다");
+
 			projectRepository.delete(project);
 			return getAllProject();
 		} catch (Exception e) {
@@ -58,14 +67,13 @@ public class ProjectService {
 	public Result<List<ProjectDto>> modifyProject(ProjectDto dto) {
 		try {
 			Project project = projectRepository.findById(Long.valueOf(dto.getId()))
-					.orElseThrow(() -> new NoSearchObjectException("프로젝트"));
+					.orElseThrow(() -> new NoSearchObjectException(DBNAME.PROJECT.getKrname()));
 			project.modify(dto);
 			if (dto.getCompanyId() != null) {
 				Company company = companyRepository.findById(Long.valueOf(dto.getCompanyId()))
-						.orElseThrow(() -> new NoSearchObjectException("회사"));
+						.orElseThrow(() -> new NoSearchObjectException(DBNAME.COMPANY.getKrname()));
 				project.setCompany(company);
 			}
-			projectRepository.delete(project);
 			return getAllProject();
 		} catch (Exception e) {
 			return new Result<List<ProjectDto>>(null, e.getMessage());
@@ -80,7 +88,25 @@ public class ProjectService {
 
 	private void checkDuplicateTitle(String title) throws DuplicateException {
 		Project duplicate = projectRepository.findByTitle(title);
-		if (duplicate == null)
+		if (duplicate != null)
 			throw new DuplicateException(title);
+	}
+
+	public Result<Map<String, List<BaseDto>>> getCreateProjectPrev() {
+		Map<String, List<BaseDto>> result = new HashMap<String, List<BaseDto>>();
+		try {
+			List<BaseDto> companyList = companyRepository.findAll().stream()
+					.map(c -> new BaseDto(c.getId().toString(), c.getName())).collect(Collectors.toList());
+			result.put("companyList", companyList);
+			List<BaseDto> statusList = new ArrayList<BaseDto>();
+			for (PSTATUS status : PSTATUS.values()) {
+				BaseDto bDto = new BaseDto(status.name(), status.getTitle());
+				statusList.add(bDto);
+			}
+			result.put("statusList", statusList);
+			return new Result<Map<String, List<BaseDto>>>(result, Constants.RESULT_SUCCESS);
+		} catch (Exception e) {
+			return new Result<Map<String, List<BaseDto>>>(null, e.getMessage());
+		}
 	}
 }
